net: error out unless net_recv_all receives all bytes

+++ src/net.h
@@ -12,7 +12,7 @@ int net_connect(const char *host, const char *port);
 int net_send_all(int fd, const void *, size_t);
 int net_send_buf(int fd, const void *, uint32_t);
 
-ssize_t net_recv_all(int fd, void *, size_t);
+int net_recv_all(int fd, void *, size_t);
 int net_recv_buf(int fd, void **, uint32_t *);
 int net_recv_static(int fd, void *, size_t);
